add `onMessage` and `onError` to `useNatsSubscription`
This way we can keep the `useNatsSubscription` relatively simple, and
it should just call some functions when a message comes in.

When we need to unmount, the active subscription will be unsubscribed.

Note: when the `onError`, `onMessage` or `subj` changes, we will setup a
new subscription as well. So make sure to always provide a stable
reference for those callback functions.