add `useIras` hook
This hook will use the `useNatsSubscription` under the hood, but will
also massage the incoming data to a data format you can work with.

The data, that was only available in the callbacks, will now also be
available in the hook (and the component that uses the hook), because we
use `useReducer` to reduce the incoming data into a proper data
structure.